crewai/code_interpreter_crew/requirements.txt
langchain-openai
langchain
open-interpreter
gradio
crewai
